Refactor test cases from monolith to smaller class files. Remove old tests.
